functions/ai_send_text_assistant_message/requirements.txt
functions-framework==3.*
google-cloud-firestore
google-cloud-storage
requests
markdown
python-dotenv
